+++ buildSrc/src/main/java/com/pedjak/gradle/plugins/dockerizedtest/DockerizedExecHandle.java
@@ -297,7 +297,7 @@ public class DockerizedExecHandle implements ExecHandle, ProcessSettings {
       while (stateIn(ExecHandleState.STARTING)) {
         LOGGER.debug("Waiting until process started: {}.", displayName);
         try {
-          if (!stateChanged.await(10, TimeUnit.MINUTES)) {
+          if (!stateChanged.await(5, TimeUnit.MINUTES)) {
             execHandleRunner.abortProcess();
             throw new RuntimeException("Giving up on " + execHandleRunner);
           }
@@ -640,7 +640,7 @@ public class DockerizedExecHandle implements ExecHandle, ProcessSettings {
           .withStdErr(true)
           .withStdIn(stdInReadStream)
           .exec(attachContainerResultCallback);
-      if (!attachContainerResultCallback.awaitStarted(2, TimeUnit.MINUTES)) {
+      if (!attachContainerResultCallback.awaitStarted(10, TimeUnit.SECONDS)) {
         LOGGER.warn("Not attached to container " + containerId + " within 10secs");
         throw new RuntimeException("Not attached to container " + containerId + " within 10secs");
       }
