+++ geode-core/src/main/java/org/apache/geode/admin/jmx/internal/AgentLauncher.java
@@ -47,9 +47,9 @@ import org.apache.geode.admin.jmx.AgentFactory;
 import org.apache.geode.distributed.internal.ClusterDistributionManager;
 import org.apache.geode.internal.ExitCode;
 import org.apache.geode.internal.net.SocketCreator;
+import org.apache.geode.internal.process.utils.OSProcess;
 import org.apache.geode.internal.util.IOUtils;
 import org.apache.geode.internal.util.JavaCommandBuilder;
-import org.apache.geode.logging.internal.OSProcess;
 import org.apache.geode.logging.internal.executors.LoggingThread;
 import org.apache.geode.logging.internal.log4j.api.LogService;
