+++ geode-core/src/main/java/org/apache/geode/distributed/internal/locks/DLockService.java
@@ -54,13 +54,13 @@ import org.apache.geode.distributed.internal.locks.DLockQueryProcessor.DLockQuer
 import org.apache.geode.distributed.internal.membership.InternalDistributedMember;
 import org.apache.geode.internal.Assert;
 import org.apache.geode.internal.logging.log4j.LogMarker;
+import org.apache.geode.internal.process.utils.OSProcess;
 import org.apache.geode.internal.serialization.DataSerializableFixedID;
 import org.apache.geode.internal.serialization.DeserializationContext;
 import org.apache.geode.internal.serialization.KnownVersion;
 import org.apache.geode.internal.serialization.SerializationContext;
 import org.apache.geode.internal.util.StopWatch;
 import org.apache.geode.internal.util.concurrent.FutureResult;
-import org.apache.geode.logging.internal.OSProcess;
 import org.apache.geode.logging.internal.log4j.api.LogService;
 import org.apache.geode.util.internal.GeodeGlossary;
