+++ geode-core/src/main/java/org/apache/geode/distributed/internal/membership/InternalDistributedMember.java
@@ -45,12 +45,12 @@ import org.apache.geode.distributed.internal.membership.api.MemberIdentifierFact
 import org.apache.geode.internal.cache.versions.VersionSource;
 import org.apache.geode.internal.inet.LocalHostUtil;
 import org.apache.geode.internal.net.SocketCreator;
+import org.apache.geode.internal.process.utils.OSProcess;
 import org.apache.geode.internal.serialization.DataSerializableFixedID;
 import org.apache.geode.internal.serialization.DeserializationContext;
 import org.apache.geode.internal.serialization.KnownVersion;
 import org.apache.geode.internal.serialization.SerializationContext;
 import org.apache.geode.internal.serialization.Version;
-import org.apache.geode.logging.internal.OSProcess;
 
 /**
  * This is the fundamental representation of a member of a GemFire distributed system.
