+++ geode-core/src/main/java/org/apache/geode/internal/ManagerInfo.java
@@ -36,7 +36,7 @@ import org.apache.geode.SystemIsRunningException;
 import org.apache.geode.UncreatedSystemException;
 import org.apache.geode.UnstartedSystemException;
 import org.apache.geode.annotations.Immutable;
-import org.apache.geode.logging.internal.OSProcess;
+import org.apache.geode.internal.process.utils.OSProcess;
 
 /**
  * Represents the information about the manager that is stored in its SystemAdmin manager VM's main
