+++ geode-core/src/main/java/org/apache/geode/internal/logging/Banner.java
@@ -42,10 +42,10 @@ import java.util.TreeMap;
 import org.apache.geode.SystemFailure;
 import org.apache.geode.internal.GemFireVersion;
 import org.apache.geode.internal.VersionDescription;
+import org.apache.geode.internal.process.utils.OSProcess;
 import org.apache.geode.internal.serialization.KnownVersion;
 import org.apache.geode.internal.util.ArgumentRedactor;
 import org.apache.geode.logging.internal.ConfigurationInfo;
-import org.apache.geode.logging.internal.OSProcess;
 import org.apache.geode.util.internal.GeodeGlossary;
 
 /**
