+++ geode-core/src/main/java/org/apache/geode/internal/logging/ManagerLogWriter.java
@@ -24,8 +24,8 @@ import java.io.PrintWriter;
 import org.apache.geode.LogWriter;
 import org.apache.geode.internal.io.MainWithChildrenRollingFileHandler;
 import org.apache.geode.internal.io.RollingFileHandler;
+import org.apache.geode.internal.process.utils.OSProcess;
 import org.apache.geode.internal.util.LogFileUtils;
-import org.apache.geode.logging.internal.OSProcess;
 import org.apache.geode.logging.internal.spi.LogConfig;
 import org.apache.geode.logging.internal.spi.LogFileDetails;
 import org.apache.geode.util.internal.GeodeGlossary;
