+++ geode-core/src/main/java/org/apache/geode/internal/logging/ManagerLogWriterFactory.java
@@ -28,9 +28,9 @@ import org.apache.logging.log4j.status.StatusLogger;
 import org.apache.geode.GemFireIOException;
 import org.apache.geode.annotations.VisibleForTesting;
 import org.apache.geode.internal.process.ProcessLauncherContext;
+import org.apache.geode.internal.process.utils.OSProcess;
 import org.apache.geode.internal.statistics.StatisticsConfig;
 import org.apache.geode.internal.util.LogFileUtils;
-import org.apache.geode.logging.internal.OSProcess;
 import org.apache.geode.logging.internal.spi.LogConfig;
 
 /**
