+++ geode-dunit/src/main/java/org/apache/geode/test/dunit/internal/ChildVM.java
@@ -20,7 +20,7 @@ import org.apache.logging.log4j.Logger;
 
 import org.apache.geode.internal.AvailablePortHelper;
 import org.apache.geode.internal.ExitCode;
-import org.apache.geode.logging.internal.OSProcess;
+import org.apache.geode.internal.process.utils.OSProcess;
 import org.apache.geode.logging.internal.log4j.api.LogService;
 import org.apache.geode.test.version.VersionManager;
