+++ geode-gfsh/src/main/java/org/apache/geode/management/internal/cli/functions/GetStackTracesFunction.java
@@ -20,7 +20,7 @@ import java.time.Instant;
 import org.apache.geode.cache.Cache;
 import org.apache.geode.cache.execute.FunctionContext;
 import org.apache.geode.internal.cache.execute.InternalFunction;
-import org.apache.geode.logging.internal.OSProcess;
+import org.apache.geode.internal.process.utils.OSProcess;
 import org.apache.geode.management.internal.cli.domain.StackTracesPerMember;
 
 public class GetStackTracesFunction implements InternalFunction<Void> {
