+++ geode-membership/src/integrationTest/java/org/apache/geode/distributed/internal/membership/MembershipDependenciesJUnitTest.java
@@ -57,6 +57,7 @@ public class MembershipDependenciesJUnitTest {
               .or(resideInAPackage("org.apache.geode.distributed.internal.tcpserver.."))
               .or(resideInAPackage("org.apache.geode.internal.inet.."))
               .or(resideInAPackage("org.apache.geode.internal.lang.."))
+              .or(resideInAPackage("org.apache.geode.internal.process.."))
               .or(resideInAPackage("org.apache.geode.annotations.."))
 
               .or(not(resideInAPackage("org.apache.geode..")))
