+++ gradle.properties
@@ -79,7 +79,7 @@ org.gradle.configureondemand = false
 org.gradle.daemon = true
 org.gradle.jvmargs = -Xmx3g
 org.gradle.parallel = true
-#org.gradle.workers.max = 3
+org.gradle.workers.max = 3
 
 org.gradle.internal.http.socketTimeout=120000
 org.gradle.internal.http.connectionTimeout=120000
