+++ settings.gradle
@@ -70,7 +70,6 @@ include 'boms:geode-client-bom'
 include 'boms:geode-all-bom'
 include 'static-analysis:pmd-rules'
 
-
 ['1.0.0-incubating',
  '1.1.0',
  '1.1.1',
@@ -102,3 +101,4 @@ buildCache {
     enabled = true
   }
 }
+
